+++ module/sheets/actor-sheet.mjs
@@ -1,0 +1,8 @@
+export function MasksActorSheetMixin(Base) {
+	return class MasksActorSheet extends Base {
+		/** @override */
+		get template() {
+			return 'modules/masks-newgeneration-unofficial/templates/sheets/actor-sheet.hbs';
+		}	
+	}
+}